[proliferate] split AnyHarness session event sink
