+++ anyharness/crates/anyharness-lib/src/acp/event_sink/normalization/mod.rs
@@ -1,0 +1,6 @@
+pub(in crate::acp::event_sink) mod file_references;
+pub(in crate::acp::event_sink) mod files;
+pub(in crate::acp::event_sink) mod meta;
+pub(in crate::acp::event_sink) mod snapshots;
+pub(in crate::acp::event_sink) mod terminals;
+pub(in crate::acp::event_sink) mod text;
